I'm finally getting around to committing the approved modification to gather_docs() I proposed to Lunar-Dev ML about a month ago.
As a refresher (or introduction for some) this is what the change is all
about.  It's an extended version of the already existing function
gather_docs().  It allows a module writer/maintainer to be able to easily
install any extra documentation (above and beyond what is already
automatically copied over) they feel would be nice to add to the module's
default document directory.

A common usage would be a line like this found at the end of a BUILD file:

gather_docs TODO INDEX Notes extras/*

Which would copy the files TODO, INDEX, Notes, and create a subdirectory
called extras and copy any files found in that directory to the new
directory under the module's default document directory.

It obeys (checks for) $GARBAGE=on for you, so you don't need to worry about
that when writing the module. The files in the file list must all reside in
or below the $SOURCE_DIRECTORY and must be referenced relative to said
directory.

There is no harm in implementing this now within the moonbase at will.  If
the user does not have the extended version of gather_docs() (and they have
verbose on) s/he will see the default documentation be coppied twice but
nothing more.