This is the commit to go along with my Dev-ML email (subject: Addition to build.lunar  common functions) sent on 18Sep05.  This adds a default_game_build and default_game_config as options to be used in BUILD files.  This commit also fixes a few minor typos.
!!!!!!
Please DO NOT use these two functions in any modules that are in the moonbase.  This 
code is being commited here so that devs can easily test the new code.  Once it has 
been tested/approved the change can then be pushed the lunar and both theedge and 
lunar can get a version bump before we can start making use of it, else stuff will 
break for the user.
!!!!!!!