[PATCH] new alias code
This new alias algorithm stores the users' choice when asked for %X
as LUNAR_ALIAS_X in the local config. This speeds up lookups later
and eliminates the need to detect aliases on the fly. We can delete
all the detection code and the user will only once ever get asked.

unsetting an alias choice is as easy as `lunar unset LUANR_ALIAS_X`