[FIX] make sure alias doesn't loop
find_depends was using a stdin | read loop which breaks anything
in that context that might read from user input. We take a small
performance hit by not using the smart awk call but instead split
up the STATE lookup which is not executed for each round. All in all
this should perform the same as the old code.

some small whitespace fixes here too.