Bugfix: download error code did not get trapped.
I had a failed moonbase delete all of my modules once, and it turns out that
the error code for file downloads is not handled properly. To make it a bit
more understandable, I added 2 state tracking variables to make sure that
we attempt to download ALL files needed, but trap errors as critical on any
one of them, so we get as much files as possible before giving up.