Overhaul BUILD file processing.
*1*
BUILD procedural change - makes C_FIFO usage in BUILD obsolete. We can
do this in the run_build() file globally for each and every BUILD file
anyway, so there's no reason to even do this at all.

We'll need to clean up moonbase entirely, but this will save us 3k occurrences
plus trimmable whitespace = well over 100k on the size of moonbase. Plus
it's easier to write.

*2*
Enable shopt -s execfail for all BUILD scripts too. We're seeing too many
undetected errors, and really should stop using && as a measure to trap
errors. Instead, with this change, we can remove all the '&&'s and in the
few occasions (not many) where errors need to be ignored, we can '|| :'
them instead, which is easy to review and check.