don't output anything unless there is progress.
when PRE_BUILD is running, the voyeur code is already running and
would print the status bar over the kconfig menu every 2 seconds.

the fix is easy: when reaching the pipe read timeout, we can just
loop and wait for more input again instead of printing the progress
bar for no reason.