build.lunar: Changing the behavior of cmake_build_target(). The instigator for this function
was kdebindings-4.6.x not picking up our $OPTS with an out of source build. Now that kde-4.7.0
is out it no longer exists with its guts being separated into the individual bindings.

I ran into a problem with kdepim-runtime-4.7.0; doing an in source build would fail because it
would not use a settings.h file and thus tank on some akonadi stuff during the make. After some
time a #akonadi an out of source build was done and all was well.

If you are realizing there are inconsistancies between these build types you are correct. Of all the
documentation I have read either should work fine with an out of source being the *prefered* method.

So this commit simply does away with the detection of that macro and defaults to an out of source. For
those modules that need an in source build the dev can simply use something like;

cmake . $OPTS blah blah.

in the BUILD.