lin: drop the 'smart' downloader code.
This code suffers horribly from the size of moonbase and can take minutes
easily when doing a lunar update, even on my i7 extreme.

Disable it for now, the penalty of having to download modules later on is
by far not as bad as this overall hit on every module install.