downloader: fix some not-well tested codepaths.
This was broken in several ways. The plugins return '2' when they
fail since more plugins need to be tried, but, we need to trap
the error code in the main loop in download_module, and not in
download_url. Taking the error code handling out fixes that problem
which caused the downloader to only try the first tarball URL.

Second problem was that the mirror inserter should append behind
all the listed URLS, not in between url[0] and url[1].