Readable temp files.
You ever wondered what information of the build process is stored
in which /tmp file? Something like lunar.lin.27767..ccOsV37CAJ
or lunar.lin.27767.octave.2p6n3jJ7ya doesn't give away too much.
I'm not sure whether or not this was intended, but the fourth
slot of the filename is set to the current module, instead
of the _purpose_ that is given as a parameter(and that
purpose seems to include the current module whenever there is one).

The patch results in filenames like
lunar.lin.27767.depends.cache.ccOsV37CAJ or
lunar.lin.27767.octave.compile-log.2p6n3jJ7ya