make `lin moonbase` do the right thing for git moonbases, etc.
This fixes a long standing bug for folks using a git moonbase. The
new code properly skips downloading a moonbase now if you `lin
moonbase`, and instead creates caches and indexes reflecting the
state of the custom moonbase, instead of the system one.

a REQUIREMENT is that you do NOT use /var/lib/lunar/moonbase anymore,
but instead do `lunar set MOONBASE /some/path/where/your/stuff/is`.
If you store a custom moonbase at the old location, it will be
overwritten.

You still need to lin moonbase to update the caches and indexes.