Added quotes around file argument of progs which read stdin
This will make the progs fail on file not found instead of handing on
stdin.
For consistency all arguments which contain a variable are quoted also
if they have a fixed part. If they only have a fixed part they are not
quoted. One exception, when the argument contains a wildcard (*) this if
not quoted, or the wildcard would fail.