Connect() no longer needed
Due to a change in the interface naming scheme as of the latest
release of systemd there's no longer a safe way to detect interface
names thus rendering this code deprecated.

Automatic dialup code has been removed as of this commit and we no
longer check if an active interface exists, lget/lin will just
timeout if there is no internet connectivity.