Added SOURCEx_UPSTREAM to rename downloaded files.
SOURCEx will specify the name as stored in SOURCE_CACHE. The optional
SOURCEx_UPSTREAM specifies the real name of the download file wich will
be renamed during download. SOURCEx_UPSTREAM will not be supported by
all download plugins and might have a different meaning for others. For
example when the SOURCEx_URL is a repository and it support branches the
SOURCEx_UPSTREAM specifies the branch name.