plugins.lunar: Clean up the code a bit, simplify things.
The previous plugins.lunar seemed to sometimes get confused about what
to do, leading to occasionally having plugins fail to be installed when
they should have been.  For instance, the dracut plugin on a fresh
install wasn't being installed, causing a fresh install to fail to boot
after installation finished.

This cleans up the code a bit to, with any luck, make that not happen as
much.