installwatch: omit fchmod & access
fchmod applies to a file descriptor, so it's just not helpful.

access is disabled in our installwatch/BUILD module,
but if you build installwatch by hand and you miss the flag,
it suddenly starts tracking all kind of files it's not supposed to.

I believe there's no harm in removing them here.