First implementation of OBSOLETE libs feature #142
The new feature KEEP_OBSOLETE_LIBS is on by default.

Libs with changed soname's will be automatically copied to
/usr/lib/lunar/OBSOLETE. This makes it safe to lin modules again, or update
modules that have not changed the lib version without ending up
with duplicated libs in OBSOLETE.

`lunar fix` will cleanup the OBSOLETE folder, but only if _ALL_ integrity checks
pass. And only if `lunar fix` is executed without module arguments, e.g. `lunar
fix module` will not clean OBSOLETE even if that module is fixed. The reason is
that multiple modules may still depend on the OBSOLETE libs.

TODO: fix so that only the LDD_CHECK need to pass for the cleanup function to
execute.