optimize.lunar: Ensure that flags are in the correct format
The daily builds everything with -Os for stage2, but -Os is a bad flag
for glibc, and -Os is either first or last in the CFLAGS, which makes
it so that bad_flags doesn't see it and doesn't remove it.

Therefore, put spaces before and after the contents of the various flags
variables so that bad_flags operates on them the right way.