fix-add-broken-depends: Ensure that it sees a string as a string
Bash has a ?: syntax which can be applied during parameter
expansion. shfmt, which contains a static analyser, identified
this bit of code as being something which could potentially
confuse the parser into thinking it's part of a ternary-operator
sequence.

So I just moved the key, with its colon in it, to a string variable
and removed all potential for confusion.