lvu depends: Significantly speed things up
Previously, `lvu depends` would grep through /var/state/lunar/depends,
once for every dependency that it came across, with a potential
combinatorial explosion for something which lots and lots of stuff
depends on (like openssl, or perl).

This gets rid of the invocations of grep, in favor of doing everything
within the same process.  It loads up an associative array with the
relevant contents of /var/state/lunar/depends, and does all of its
operations in RAM instead of having to go through the Lunar depends
state file multiple times (I tried `lvu depends perl` in strace, and it
exec'ed /usr/bin/grep over 14,000 times!).