files.lunar: Halfway fix for the loop detector.
First of all, user "error_message" so that the error messages go to
standard error (duh) instead of polluting the input of whatever's
trying to call it.

Second of all, if standard out is a terminal, then ask questions. If
not, just soldier on. As the comment says, this is probably not the
ideal way to proceed.